fastapi_app/requirements_fastapi.txt
fastapi[all]
loguru
python-dotenv
python_magic
redis
tqdm
flower
gunicorn
celery
aiofiles
aioredis
